==> notebooks/01_data_cleaning_merge.ipynb <==
## Overview

- Read the downloaded datasets
- Merge the individual datasets into one DataFrame
- Clean the DataFrame
- Save the DataFrame as a new file for exploratory data analysis

## Dataset

This study utilises data from the August 2021 - August 2023 cycle of National Health and Nutrition Examination Survey (NHANES) conducted by Centers for Disease Control and Prevention (CDC).

[NHANES Aug 2021-Aug 2023 Cycle] (https://wwwn.cdc.gov/nchs/nhanes/continuousnhanes/default.aspx?Cycle=2021-2023)

A few datasets of interest were selected for the purpose of analysis.

Demographic Data
- DEMO_L.xpt : Demographic Variables and Sample Weights

Examination Data
- BMX_L.xpt : Body Measures

Laboratory Data
- HDL_L.xpt : Cholesterol - High-Density Lipoprotein
- GHB_L.xpt : Glycohemoglobin
- GLU_L.xpt : Plasma Fasting Glucose

Questionnaire Data
- ALQ_L.xpt : Alcohol Use
- BPQ_L.xpt : Blood Pressure & Cholesterol
- DIQ_L.xpt : Diabetes
- SMQ_L.xpt : Smoking - Cigarette Use



import pandas as pd
from pathlib import Path

# Set the path to data directory
data_dir= Path("../data/raw")

# Create a file map for better readability
files = {
    "bmi": "BMX_L.xpt",
    "blood_pressure": "BPQ_L.xpt",
    "demographics": "DEMO_L.xpt",
    "diabetes": "DIQ_L.xpt",
    "glycohemoglobin": "GHB_L.xpt",
    "fasting_plasma_glucose": "GLU_L.xpt",
    "high_density_lipoprotein": "HDL_L.xpt",
    "smoking": "SMQ_L.xpt"
}

datasets = {}
total_files_count = len(files)
current_count = 1
# Load each dataset into a pandas DataFrame
for key, filename in files.items():
    path = data_dir / filename
    print(f"File {current_count}/{total_files_count} Loading {filename} into DataFrame...")
    datasets[key] = pd.read_sas(path, format='xport')
    current_count += 1


# Read each dataset and print the first few rows to observe what the data looks like and identify any common features
datasets["bmi"].head()

datasets["blood_pressure"].head()

datasets["demographics"].head()

datasets["diabetes"].head()

datasets["glycohemoglobin"].head()

datasets["fasting_plasma_glucose"].head()

datasets["high_density_lipoprotein"].head()

datasets["smoking"].head()

# Merge the datasets on "SEQN" into a single DataFrame using Python reduce() and pandas merge()
from functools import reduce

df = reduce(
    lambda left, right: pd.merge(left, right, on="SEQN", how="outer"),
    datasets.values()
)

# Show information about the merged DataFrame
pd.set_option('display.max_rows', None)

df.info()

df.shape

# Show the first 5 rows of the merged DataFrame
pd.set_option('display.max_columns', None)
pd.set_option('display.width', None)

df.head()

The following features are deemed relevant to the problem and will be used for analysis:
- BMXBMI [Continuous] : Body Mass Index (kg/m^2)
- BPQ020 [Binary] : History of high blood pressure (0=N, 1=Y)
- RIAGENDR [Binary] : Gender (0=F, 1=M)
- RIDAGEYR [Discrete] : Age at screening (years)
- DIQ010 [Multiclass] : Diabetic (Yes, No, Borderline)
- LBXGH [Continuous] : Glycohemoglobin level (%)
- LBDGLUSI [Continuous] : Fasting Plasma Glucose (mmol/L)
- LBDHDDSI [Continuous] : Direct HDL-Cholesterol (mmol/L)
- SMQ020 [Binary] : Smoked at least 100 cigarettes in lifetime
- SMQ040 [Multiclass] : Do you now smoke cigarettes (Every day, Some day, None at all)

# Selecting relevant columns for analysis
keep_cols = ["BMXBMI", "BPQ020", "RIAGENDR", "RIDAGEYR", "DIQ010", "LBXGH", "LBDGLUSI", "LBDHDDSI", "SMQ020", "SMQ040"]

df = df[keep_cols]
df.info()

# Rename columns for better readability
df = df.rename(columns={
    "BMXBMI": "bmi",
    "BPQ020": "high_blood_pressure",
    "RIAGENDR": "gender",
    "RIDAGEYR": "age",
    "DIQ010": "diabetes",
    "LBXGH": "hba1c_percentage",
    "LBDGLUSI": "fpg_mmol_L",
    "LBDHDDSI": "hdl_mmol_L",
    "SMQ020": "lifetime_100_cigs_smoked",
    "SMQ040": "current_smoking_freq"
})

# Identify the number of missing values in each column
df.isnull().sum()

# Value counts for the "SMQ020" column (smoking status)
# df["lifetime_100_cigs_smoked"].value_counts()

# Verify that all rows with SMQ020 == 2.0 had missing values, hence no SMQ020 = 2.0 present in the resulting DataFrame
# sum(df[df["lifetime_100_cigs_smoked"] == 2.0].isnull().sum(axis=1).value_counts())

It is found that all rows with SMQ020 == 2.0 had at least one missing value. Hence the resulting DataFrame after dropping rows with missing values only has one unique value for SMQ020 (1.0).

This study will focus on participants who are aged 18 and above. Hence, all participants with age < 18 are removed from the DataFrame.

# Remove rows with missing or unwanted values (coded as "Refused" or "Don't know")
df = df.dropna()
df = df[(df["high_blood_pressure"] != 9.0)]
df = df[(df["age"] >= 18.0)]

# Resultant DataFrame info after cleaning
df.info()

# Define a function to display unique values for each column in the DataFrame
def display_unique_valeus(df):
    unique_values = {}
    for col in df.columns:
        unique_values[col] = df[col].unique().tolist()
    return unique_values

# Display unique values for each column in the DataFrame and their counts
unique_col_values = display_unique_valeus(df)
for col, vals in unique_col_values.items():
    print(f"{col} ({len(vals)}): {vals}")

# Replace gender values to 0=F, 1=M
df["gender"] = df["gender"].replace(2.0, 0.0)

# Replace high_blood_pressure values to 0 = No, 1 = Yes
df["high_blood_pressure"] = df["high_blood_pressure"].replace(2.0, 0.0)

# Convert categorical columns to integers
cols_to_int = ["high_blood_pressure", "gender", "age", "diabetes", "lifetime_100_cigs_smoked", "current_smoking_freq"]

for col in cols_to_int:
    df[col] = df[col].astype(int)

# Save the cleaned DataFrame to a CSV file
df.to_csv("../data/interim/nhanes_data_interim.csv", index=False)
